### 2023-07-21 #5674 @editorialbot
[PRE REVIEW]: BlackBIRDS: Black-Box Inference foR Differentiable Simulators
**Submitting author:** @arnauqb (<a href="http://orcid.org/0000-0001-5055-9863">Arnau Quera-Bofarull</a>)
**Repository:** https://github.com/arnauqb/blackbirds
**Branch with paper.md** (empty if default branch):
**Version:** v1.0
