+++ blackbirds/losses.py
@@ -3,6 +3,7 @@ from blackbirds.simulate import simulate_and_observe_model
 
 import torch
 
+
 class SingleOutput_SimulateAndMSELoss:
 
     """
@@ -14,12 +15,7 @@ class SingleOutput_SimulateAndMSELoss:
     - `gradient_horizon`: Specifies the gradient horizon to use. None implies infinite horizon.
     """
 
-    def __init__(
-        self, 
-        model: Model, 
-        gradient_horizon: int | None = None
-    ):
-
+    def __init__(self, model: Model, gradient_horizon: int | None = None):
         self.loss = torch.nn.MSELoss()
         self.model = model
         self.gradient_horizon = gradient_horizon
@@ -29,20 +25,12 @@ class SingleOutput_SimulateAndMSELoss:
         theta: torch.Tensor,
         y: torch.Tensor,
     ):
-
-        x = simulate_and_observe_model(
-            self.model, 
-            theta, 
-            self.gradient_horizon
-        )[0]
+        x = simulate_and_observe_model(self.model, theta, self.gradient_horizon)[0]
         return self.loss(x, y)
 
-class UnivariateMMDLoss:
-    def __init__(
-        self, 
-        y: torch.Tensor
-    ):
 
+class UnivariateMMDLoss:
+    def __init__(self, y: torch.Tensor):
         """
         Computes MMD between data y and simulated output x (to be passed during call).
 
@@ -51,51 +39,58 @@ class UnivariateMMDLoss:
 
         assert isinstance(y, torch.Tensor), "y is assumed to be a torch.Tensor here"
         try:
-            assert len(y.shape) == 1, "This class assumes y is a single univariate time series"
+            assert (
+                len(y.shape) == 1
+            ), "This class assumes y is a single univariate time series"
         except AssertionError:
-            assert len(y.shape) == 2, "If not a 1D Tensor, y must be at most 2D of shape (1, T)"
-            assert y.shape[1] == 1, "This class assumes y is a single univariate time series. This appears to be a batch of data."
+            assert (
+                len(y.shape) == 2
+            ), "If not a 1D Tensor, y must be at most 2D of shape (1, T)"
+            assert (
+                y.shape[1] == 1
+            ), "This class assumes y is a single univariate time series. This appears to be a batch of data."
             y = y.reshape(-1)
 
         self.y = y
-        self.y_matrix = self.y.reshape(1,-1,1)
+        self.y_matrix = self.y.reshape(1, -1, 1)
         yy = torch.cdist(self.y_matrix, self.y_matrix)
         yy_sqrd = torch.pow(yy, 2)
         self.y_sigma = torch.median(yy_sqrd)
         ny = self.y.shape[0]
-        self.kyy = (
-            torch.exp( 
-                -yy_sqrd / self.y_sigma 
-            ) 
-            - torch.eye(ny)
-        ).sum() / (ny * (ny - 1))
-        
+        self.kyy = (torch.exp(-yy_sqrd / self.y_sigma) - torch.eye(ny)).sum() / (
+            ny * (ny - 1)
+        )
+        print(self.kyy)
+
     def __call__(
-        self, 
+        self,
         x: torch.Tensor,
     ):
-
         assert isinstance(x, torch.Tensor), "x is assumed to be a torch.Tensor here"
         try:
-            assert len(x.shape) == 1, "This class assumes x is a single univariate time series"
+            assert (
+                len(x.shape) == 1
+            ), "This class assumes x is a single univariate time series"
         except AssertionError:
[…]
