### 2023-07-25 5343c5e
change name to test mmd loss
