+++ blackbirds/infer/vi.py
@@ -99,7 +99,9 @@ def _differentiate_loss_pathwise(parameters, jacobians):
     device = parameters.device
     to_diff = torch.zeros(1, device=device)
     for i in range(len(jacobians)):
-        to_diff += torch.dot(jacobians[i].to(device), parameters[i, :])
+        # Note: It is necessary to use F64 here otherwise it can cause nan due to overflow...
+        #to_diff += torch.dot(jacobians[i].to(device).to(torch.float64), parameters[i, :].to(torch.float64))
+        to_diff += torch.dot(jacobians[i].to(device).to(torch.float), parameters[i, :].to(torch.float))
     to_diff = to_diff / len(jacobians)
     to_diff.backward()
 
@@ -111,7 +113,6 @@ def compute_loss_and_jacobian_pathwise(
     observed_outputs: list[torch.Tensor],
     diff_mode: str = "reverse",
     jacobian_chunk_size: int | None = None,
-    gradient_horizon: int = 0,
     device: str = "cpu",
 ):
     r"""Computes the loss and the jacobian of the loss for each sample using a differentiable simulator. That is, we compute
@@ -137,7 +138,6 @@ def compute_loss_and_jacobian_pathwise(
     - `observed_outputs`: observed outputs
     - `diff_mode`: differentiation mode can be "reverse" or "forward"
     - `jacobian_chunk_size`: chunk size for the Jacobian computation (set None to get maximum chunk size)
-    - `gradient_horizon`: horizon for the gradient computation
     - `device`: device to use for the computation
     """
     # sample parameters and scatter them across devices
@@ -171,7 +171,7 @@ def compute_loss_and_jacobian_pathwise(
         if torch.isnan(loss_i) or torch.isnan(jacobian).any():
             continue
         loss += loss_i
-        jacobians_per_rank.append(torch.tensor(jacobian.cpu().numpy(), dtype=torch.float))
+        jacobians_per_rank.append(torch.tensor(jacobian.cpu().numpy()))
         indices_per_rank.append(i)
     # gather the jacobians and parameters from all ranks
     if mpi_size > 1:
@@ -280,7 +280,6 @@ def compute_and_differentiate_loss(
     diff_mode: str = "reverse",
     gradient_estimation_method: str = "pathwise",
     jacobian_chunk_size: int | None = None,
-    gradient_horizon: int = 0,
     device: str = "cpu",
 ):
     r"""Computes and differentiates the loss according to the chosen gradient estimation method
@@ -294,7 +293,6 @@ def compute_and_differentiate_loss(
     - `diff_mode`: differentiation mode can be "reverse" or "forward"
     - `gradient_estimation_method`: gradient estimation method can be "pathwise" or "score"
     - `jacobian_chunk_size`: chunk size for the Jacobian computation (set None to get maximum chunk size)
-    - `gradient_horizon`: horizon for the gradient computation
     - `device`: device to use for the computation
     """
     if gradient_estimation_method == "pathwise":
@@ -310,7 +308,6 @@ def compute_and_differentiate_loss(
             diff_mode=diff_mode,
             device=device,
             jacobian_chunk_size=jacobian_chunk_size,
-            gradient_horizon=gradient_horizon,
         )
         if mpi_rank == 0:
             _differentiate_loss_pathwise(parameters, jacobians)
@@ -348,7 +345,6 @@ class VI:
     - `diff_mode`: The differentiation mode to use. Can be either 'reverse' or 'forward'.
     - `gradient_estimation_method`: The method to use for estimating the gradients of the loss. Can be either 'pathwise' or 'score'.
     - `jacobian_chunk_size` : The number of rows computed at a time for the model Jacobian. Set to None to compute the full Jacobian at once.
[…]
