### 2023-07-25 bc991eb
mmd loss on gpu
