### 2023-07-26 cbd6e3c
progress bar only for rank 0
