### 2023-07-26 962ecf0
adapted for gpu
