+++ blackbirds/networks.py
@@ -11,16 +11,17 @@ class MLP(nn.Module):
     def __init__(self, 
                  input_dim=64,
                  hidden_dims=[32, 32],
-                 output_dim=1):
+                 output_dim=1,
+                 device="cpu"):
         
         super().__init__()
         
         self.relu = nn.ReLU()
-        self._layers = nn.Sequential(nn.Linear(input_dim, hidden_dims[0]), self.relu)
+        self._layers = nn.Sequential(nn.Linear(input_dim, hidden_dims[0], device=device), self.relu)
         for i in range(len(hidden_dims) - 1):
-            self._layers.append(nn.Linear(hidden_dims[i], hidden_dims[i+1]))
+            self._layers.append(nn.Linear(hidden_dims[i], hidden_dims[i+1], device=device))
             self._layers.append(self.relu)
-        self._layers.append(nn.Linear(hidden_dims[-1], output_dim))
+        self._layers.append(nn.Linear(hidden_dims[-1], output_dim, device=device))
 
     def forward(self, x):
 
@@ -40,16 +41,17 @@ class RNN(nn.Module):
                  num_layers=1,
                  final_ff=nn.Identity(),
                  nonlinearity='tanh',
-                 flavour='gru'):
+                 flavour='gru',
+                 device="cpu"):
         
         super().__init__()
 
         if flavour == 'gru':
             self._rnn = nn.GRU(input_size, hidden_size, num_layers,
-                               batch_first=True)
+                               batch_first=True, device=device)
         else:
             self._rnn = nn.RNN(input_size, hidden_size, num_layers, nonlinearity=nonlinearity,
-                               batch_first=True)
+                               batch_first=True, device=device)
         self._fff = final_ff
         self._rnn_n_pars = count_pars(self._rnn)
         self._fff_n_pars = count_pars(self._fff)
@@ -73,7 +75,8 @@ class CNN(nn.Module):
                  hidden_layer_channels=3,
                  conv_kernel_size=4,
                  pool_kernel_size=2,
-                 final_ff=[32, 16]):
+                 final_ff=[32, 16],
+                 device="cpu"):
 
         assert len(final_ff) == 2
 
@@ -86,16 +89,16 @@ class CNN(nn.Module):
 
         self.conv1 = nn.Conv2d(n_channels, 
                                hidden_layer_channels, 
-                               conv_kernel_size)
+                               conv_kernel_size, device=device)
         self.pool = nn.MaxPool2d(pool_kernel_size,
                                  pool_kernel_size)
         #self.conv2 = nn.Conv2d(hidden_layer_channels,
         #                       2,
         #                       conv_kernel_size)
         self.fc1 = nn.Linear(hidden_layer_channels * pool1_out_dim**2,#(pool2_out_dim)**2, 
-                             final_ff[0])
+                             final_ff[0], device=device)
         self.fc2 = nn.Linear(final_ff[0],
-                             final_ff[1])
+                             final_ff[1], device=device)
         self.relu = nn.ReLU()
 
     def forward(self, x):
