### 2023-07-29 5e57aab
remove asserts
