### 2023-08-09 #5674 @rkurchin
👋 @lcao11 and @janfb, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
