### 2023-08-11 ca79001
fixed nan gradient in mmd loss
