### 2023-08-11 bdf7bae
mmd fix device for gpu
