### 2023-08-13 aaf40cf
adapted univariate mmd loss to gpu
