### 2023-08-13 f0c0635
avoid pickling through mpi4py
