### 2023-08-22 ddb3877
make loss a float in score gradient estimation
