### 2023-08-30 arnauqb/blackbirds#42 @rajeshrinet
Multi-GPU parallelisation support with MPI
@arnauqb this issue is in reference to JOSS review: https://github.com/openjournals/joss-reviews/issues/5776

The submitted paper claims a feature on "Multi-GPU parallelisation support with MPI".

Is there an example of this in the repository? In particular, it would be nice to have a plot (either in the paper or on the repo) which shows benchmarks for parallelisation with details of the system used.
