### 2023-08-30 arnauqb/blackbirds#43 @rajeshrinet
Installation and example from README
@arnauqb this issue is in reference to JOSS review: https://github.com/openjournals/joss-reviews/issues/5776

I installed the software on my computer using `pip install git+https://github.com/arnauqb/blackbirds`

It worked. But I can not execute the example given in the readme. Here is the error message:

```
Cell In[1], line 4
      1 import torch
      3 from blackbirds.models.random_walk import RandomWalk
----> 4 from blackbirds.infer.vi import VI
      5 from blackbirds.posterior_estimators import TrainableGaussian
      6 from blackbirds.simulate import simulate_and_observe_model

File <frozen importlib._bootstrap>:1007, in _find_and_load(name, import_)

File <frozen importlib._bootstrap>:986, in _find_and_load_unlocked(name, import_)

File <frozen importlib._bootstrap>:664, in _load_unlocked(spec)

File <frozen importlib._bootstrap>:627, in _load_backward_compatible(spec)

File <frozen zipimport>:259, in load_module(self, fullname)

File ~/software/anaconda/lib/python3.9/site-packages/blackbirds-1.0.0-py3.9.egg/blackbirds/infer/__init__.py:2
      1 from .smd import SMD
----> 2 from .vi import VI
      3 from .mcmc import MALA, MCMC

File <frozen importlib._bootstrap>:1007, in _find_and_load(name, import_)

File <frozen importlib._bootstrap>:986, in _find_and_load_unlocked(name, import_)

File <frozen importlib._bootstrap>:664, in _load_unlocked(spec)

File <frozen importlib._bootstrap>:627, in _load_backward_compatible(spec)

File <frozen zipimport>:259, in load_module(self, fullname)

File ~/software/anaconda/lib/python3.9/site-packages/blackbirds-1.0.0-py3.9.egg/blackbirds/infer/vi.py:113
    103     to_diff = to_diff / len(jacobians)
    104     to_diff.backward()
    107 def compute_loss_and_jacobian_pathwise(
    108     loss_fn: Callable,
    109     posterior_estimator: Callable,
    110     n_samples: int,
    111     observed_outputs: list[torch.Tensor],
    112     diff_mode: str = "reverse",
--> 113     jacobian_chunk_size: int | None = None,
    114     gradient_horizon: int = 0,
    115     device: str = "cpu",
    116 ):
    117     r"""Computes the loss and the jacobian of the loss for each sample using a differentiable simulator. That is, we compute
    118
    119     $$
   (...)
    141     - `device`: device to use for the computation
    142     """
    143     # sample parameters and scatter them across devices

TypeError: unsupported operand type(s) for |: 'type' and 'NoneType'
```

Am I missing something?
