### 2023-09-10 arnauqb/blackbirds#44 @marvinschmitt
Extend API documentation
@arnauqb This issue is in reference to the JOSS Review https://github.com/openjournals/joss-reviews/issues/5776

Currently, the documentation website contains the API documentation to the submodules `infer` and `models`. Please also add the API documentation to all remaining functionality (e.g., `blackbirds.simulate`) to the docs.

Furthermore, there is no documentation for some code, such as the entirety of `blackbirds.posterior_estimators`. Please write docstrings and add that to the docs as well.

Thank you!
