### 2023-09-10 arnauqb/blackbirds#45 @marvinschmitt
adjust requirements.txt
@arnauqb This issue is in reference to the JOSS review https://github.com/openjournals/joss-reviews/issues/5776

Please consider adjusting your [`requirements.txt`](https://github.com/arnauqb/blackbirds/blob/main/requirements.txt) to state `>=` instead of `==`. This decreases the likelihood of unresolvable environment requirements.

This is an optional request. If you have good reasons to stick with `==`, please let me know.
