### 2023-09-10 #5776 @marvinschmitt
A few comments on the paper itself:

- The in-text citations within parentheses are incorrectly formatted. It's currently `[...] simulation models (see e.g. Bayedin et al. (2020))` and should be `[...] simulation models (see e.g., Bayedin et al., 2020)`.
- A tiny pet peeve of mine: `e.g.` should be followed by a comma. Just mentioning it because it occurred in the citation example from above.
- Please add curly braces around `{B}ayesian` in the `.bib` file for proper capitalization.
- The sentence *[Our package is used] in forthcoming publications yet to be announced publicly* doesn't age well. I would suggest removing it altogether.
- You could consider adding the `BayesFlow` package for amortized Bayesian workflows (focus on simulation-based inference) to the list of related software. I'm one of the authors of that package, so this is just a suggestion and strictly optional.

```
@misc{radev2023bayesflow,
  title = {BayesFlow: Amortized Bayesian Workflows With Neural Networks},
  author = {Stefan T Radev and Marvin Schmitt and Lukas Schumacher and Lasse Elsem\"{u}ller and Valentin Pratz and Yannik Sch\"{a}lte and Ullrich K\"{o}the and Paul-Christian B\"{u}rkner},
  year = {2023},
  publisher= {arXiv},
  url={https://arxiv.org/abs/2306.16015}
}
```
