### 2023-09-19 b58ce0f
adapting gpu parallelisation example
