### 2023-09-19 c91a9d6
update multi-gpu script
