### 2023-09-19 d86d855
Use old Union for union types, test python 3.9 and 3.10 in gh actions
