### 2023-09-19 arnauqb/blackbirds#43 @arnauqb
This syntax was introduced in Python 3.10 (https://peps.python.org/pep-0604/) so you probably have an older Python version. I will revert to using the `Union` function from `typing` since Python 3.10 is probably a bit too new to enforce it.
