+++ blackbirds/mpi_setup.py
@@ -4,7 +4,7 @@ try:
     mpi_comm = MPI.COMM_WORLD
     mpi_rank = mpi_comm.Get_rank()
     mpi_size = mpi_comm.Get_size()
-except:
+except ImportError:
     mpi_comm = None
     mpi_rank = 0
     mpi_size = 1
