+++ docs/api/posterior_estimators.md
@@ -1,0 +1,3 @@
+# Example posterior estimators 
+
+:::blackbirds.posterior_estimators
