### 2023-09-19 319821c
added instructions on how to run tests and link to contributing:
