### 2023-09-19 arnauqb/blackbirds#44 @arnauqb
Should be addressed in https://github.com/arnauqb/blackbirds/pull/46/commits/6ef774955981b1606f51dc8e5965c56fb0bd60ad as part of https://github.com/arnauqb/blackbirds/pull/46
