### 2023-09-19 44f3bc1
added gpu details on parallelization example
