+++ examples/gpu_parallelization.md
@@ -122,3 +122,5 @@ mpirun -np 5 python examples/variational_inference/05-gpu_parallelisation.py --d
 In this case, we will use 5 GPUs, indexed by the ids we pass. Each GPU will execute 4 evaluations of the model for a total of 20. We can plot the scaling with number of GPUs for this particular model:
 
 ![](./gpu_scaling.png)
+
+The GPUs used are 8x Tesla V100 32GB.
