### 2023-09-19 arnauqb/blackbirds#42 @arnauqb
Added this example: https://github.com/arnauqb/blackbirds/blob/joss_review/examples/gpu_parallelization.md with a basic benchmark
