### 2023-09-19 4d43b90
added bayesflow and corrected references in paper
