+++ paper/paper.bib
@@ -76,20 +76,22 @@ numpages = {12}
   number={5},
   pages={804--815},
   year={2009},
+  doi={10.1016/j.jempfin.2009.06.006},
   publisher={Elsevier}
 }
 
 @article{mala,
-  title={Exponential convergence of Langevin distributions and their discrete approximations},
+  title={Exponential convergence of {L}angevin distributions and their discrete approximations},
   author={Roberts, Gareth O and Tweedie, Richard L},
   journal={Bernoulli},
   pages={341--363},
   year={1996},
+  doi={10.2307/3318418},
   publisher={JSTOR}
 }
 
 @article{hmc,
-  title={Hybrid monte carlo},
+  title={Hybrid {M}onte {C}arlo},
   author={Duane, Simon and Kennedy, Anthony D and Pendleton, Brian J and Roweth, Duncan},
   journal={Physics letters B},
   volume={195},
@@ -107,11 +109,12 @@ numpages = {12}
   number={5},
   pages={1103},
   year={2016},
+  doi={10.1111/rssb.12158},
   publisher={Wiley-Blackwell}
 }
 
 @article{gvi,
-  title={{An optimization-centric view on Bayes’ rule: Reviewing and generalizing variational inference}},
+  title={{An optimization-centric view on {B}ayes’ rule: Reviewing and generalizing variational inference}},
   author={Knoblauch, Jeremias and Jewson, Jack and Damoulas, Theodoros},
   journal={Journal of Machine Learning Research},
   volume={23},
@@ -121,7 +124,7 @@ numpages = {12}
 }
 
 @article{ai4abm,
-  title={Bayesian calibration of differentiable agent-based models},
+  title={{B}ayesian calibration of differentiable agent-based models},
   author={Quera-Bofarull, Arnau and Chopra, Ayush and Calinescu, Anisoara and Wooldridge, Michael and Dyer, Joel},
   journal={ICLR Workshop on AI for Agent-based Modelling},
   year={2023}
@@ -149,10 +152,18 @@ numpages = {12}
 
 @article{blit, doi = {10.21105/joss.04622}, url = {https://doi.org/10.21105/joss.04622}, year = {2022}, publisher = {The Open Journal}, volume = {7}, number = {79}, pages = {4622}, author = {Marco Benedetti and Gennaro Catapano and Francesco De Sclavis and Marco Favorito and Aldo Glielmo and Davide Magnanimi and Antonio Muci}, title = {Black-it: A Ready-to-Use and Easy-to-Extend Calibration Kit for Agent-based Models}, journal = {Journal of Open Source Software} }
 
-@article{pyvbmc, doi = {10.21105/joss.05428}, url = {https://doi.org/10.21105/joss.05428}, year = {2023}, publisher = {The Open Journal}, volume = {8}, number = {86}, pages = {5428}, author = {Bobby Huggins and Chengkun Li and Marlon Tobaben and Mikko J. Aarnos and Luigi Acerbi}, title = {PyVBMC: Efficient Bayesian inference in Python}, journal = {Journal of Open Source Software} }
+@article{pyvbmc, 
+    doi = {10.21105/joss.05428}, 
+    url = {https://doi.org/10.21105/joss.05428}, 
+    year = {2023}, 
+    publisher = {The Open Journal}, 
+    volume = {8}, number = {86}, pages = {5428}, 
+    author = {Bobby Huggins and Chengkun Li and Marlon Tobaben and Mikko J. Aarnos and Luigi Acerbi}, 
+    title = {PyVBMC: Efficient {B}ayesian inference in Python}, journal = {Journal of Open Source Software}
+}
 
 @article{abcpy,
[…]
