### 2023-09-19 7c0295c
delete new mmd loss
