### 2023-09-19 9c680ab
Update gpu_parallelization.md
