### 2023-09-19 64d7a2b
removing brackets from citations inside brackets
