+++ paper/paper.bib
@@ -188,7 +188,7 @@ numpages = {12}
 }
 
 @misc{radev2023bayesflow,
-  title = {BayesFlow: Amortized Bayesian Workflows With Neural Networks},
+  title = {BayesFlow: Amortized {B}ayesian Workflows With Neural Networks},
   author = {Stefan T Radev and Marvin Schmitt and Lukas Schumacher and Lasse Elsem\"{u}ller and Valentin Pratz and Yannik Sch\"{a}lte and Ullrich K\"{o}the and Paul-Christian B\"{u}rkner},
   year = {2023},
   publisher= {arXiv},
