### 2023-09-19 b43c088
upgrade docs to python 3.10 to solve recursion problem?
