+++ .github/workflows/docs.yml
@@ -12,7 +12,7 @@ jobs:
   build:
     strategy:
       matrix:
-        python-version: [ 3.9 ]
+        python-version: [ 3.10 ]
         os: [ ubuntu-latest ]
     runs-on: ${{ matrix.os }}
     steps:
