### 2023-09-20 #5776 @rajeshrinet
@rkurchin  I have completed my first review of the code and the paper. The software looks very good. The paper is well written with a clear statement of need and state of the field.

I found a few problems with installation, but [@](https://github.com/rajeshrinet)[arnauqb](https://github.com/arnauqb) addressed those.

I have made a few suggestions via issues on the code repository, which the authors have handled.

Some comments (optional) for the authors to consider:

* High-level summary in the paper: the current summary of the paper is clear enough for potential software users but not for a diverse, non-specialist audience.   Similarly, there needs to be more description (in the form of comments) in the example files on GitHub. I suggest adding a README in the `examples` folder and some more comments in the representative examples.

* State of the field: the paper compares with other commonly used packages. In addition, I want to bring to the attention, the PyRoss library: https://github.com/rajeshrinet/pyross. I am one of the team members who developed PyRoss. PyRoss allows full Bayesian inference on any user-defined compartment epidemiological models through Python dictionaries. The inherent method -  derived from functional limit theorems for Markov population process - is not limited to compartment epidemiological models.
