### 2023-09-25 #5776 @marvinschmitt
Thank you for addressing my comments @arnauqb. Based on the edits, I have updated my review checklist.

BTW: The BayesFlow paper has just been published in JOSS last week, so you could directly update the reference.

```
@article{Radev_BayesFlow_Amortized_Bayesian_2023,
    author = {Radev, Stefan T. and Schmitt, Marvin and Schumacher, Lukas and Elsemüller, Lasse and Pratz, Valentin and Schälte, Yannik and Köthe, Ullrich and Bürkner, Paul-Christian},
    doi = {10.21105/joss.05702},
    journal = {Journal of Open Source Software},
    month = sep,
    number = {89},
    pages = {5702},
    title = {{BayesFlow: Amortized Bayesian Workflows With Neural Networks}},
    url = {https://joss.theoj.org/papers/10.21105/joss.05702},
    volume = {8},
    year = {2023}
}
```
