### 2023-09-25 6f2a8ab
added examples readme, updated bayesflow reference
