+++ paper/paper.bib
@@ -187,10 +187,15 @@ numpages = {12}
   url = {https://doi.org/10.21105/joss.04304},
 }
 
-@misc{radev2023bayesflow,
-  title = {BayesFlow: Amortized {B}ayesian Workflows With Neural Networks},
-  author = {Stefan T Radev and Marvin Schmitt and Lukas Schumacher and Lasse Elsem\"{u}ller and Valentin Pratz and Yannik Sch\"{a}lte and Ullrich K\"{o}the and Paul-Christian B\"{u}rkner},
-  year = {2023},
-  publisher= {arXiv},
-  url={https://arxiv.org/abs/2306.16015}
+@article{radev2023bayesflow,
+    author = {Radev, Stefan T. and Schmitt, Marvin and Schumacher, Lukas and Elsemüller, Lasse and Pratz, Valentin and Schälte, Yannik and Köthe, Ullrich and Bürkner, Paul-Christian},
+    doi = {10.21105/joss.05702},
+    journal = {Journal of Open Source Software},
+    month = sep,
+    number = {89},
+    pages = {5702},
+    title = {{BayesFlow: Amortized Bayesian Workflows With Neural Networks}},
+    url = {https://joss.theoj.org/papers/10.21105/joss.05702},
+    volume = {8},
+    year = {2023}
 }
