### 2023-09-25 #5776 @arnauqb
I've updated the BayesFlow reference and added an additional README to the examples folder. Thank you very much for your helpful comments and additions.

I believe this concludes the review @rkurchin ?
