### 2023-09-25 46f6883
Update setup.py
update to v1.2
