### 2023-09-25 #5776 @arnauqb
* New version with all implemented changes: v1.2 (https://github.com/arnauqb/blackbirds/releases/tag/v1.2)
* Zenodo DOI: 10.5281/zenodo.8377044
