### 2023-09-25 #5776 @rkurchin
Editorial comments:
* line 19: "means to" somehow sounds weird to me and it feels like it should be "means of" although I can't explain why (this should be considered an optional change)
* second page: be consistent about whether equations in the middle of a sentence have commas after them (I lean towards including them, i.e. adding them in equations (2) and (3) to be consistent with (1))

I'm pretty sure this is the fewest editorial comments I've ever given! 👏

Conceptual question: Is it currently the case that everything needs to be in PyTorch? For example, could a model be in JAX, or even in some other AD-able language like Julia? It would be helpful to clarify this, and if it's not currently possible, it's obviously beyond the scope of this review but I would selfishly be thrilled if this could be possible 😁
