### 2023-09-25 #5776 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
