### 2023-09-29 760d660
paper published at JOSS! added citation :rocket:
