### 2025-01-28 #5776 @homerjed
👋 @arnauqb, the `editorialbot` has listed you as a potential reviewer (see here https://github.com/openjournals/joss-reviews/issues/7634) for my software submission (https://github.com/homerjed/sbgm). If you are available, would you consider reviewing? Cheers!
